Inns are now booths
