+++ DragonQuestino/game_data.c
@@ -13954,8 +13954,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
    if ( id == TILEMAP_OVERWORLD_ID ) tileMap->gameFlags->doors = ( 0xFFFF0000 | ( tileMap->gameFlags->doors & 0xFFFF ) );
 
    Random_Seed();
-   tileMap->innTileIndex = -1;
-   tileMap->innPrice = 0;
    tileMap->boothCount = 0;
 
    switch( id )
@@ -13969,8 +13967,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 140;
          tileMap->tilesY = 135;
          tileMap->portalCount = 14;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 1272;
          tileMap->portals[0].destinationTileMapIndex = 7;
          tileMap->portals[0].destinationTileIndex = 331;
@@ -18181,8 +18177,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 213;
          tileMap->portals[0].destinationTileMapIndex = 3;
          tileMap->portals[0].destinationTileIndex = 264;
@@ -18302,8 +18296,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 125;
          tileMap->portals[0].destinationTileMapIndex = 3;
          tileMap->portals[0].destinationTileIndex = 990;
@@ -18382,8 +18374,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 32;
          tileMap->portalCount = 64;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 995;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 7193;
@@ -19164,11 +19154,11 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 32;
          tileMap->portalCount = 15;
-         tileMap->innTileIndex = 714;
-         tileMap->innPrice = 6;
-         tileMap->boothCount = 1;
-         tileMap->booths[0].id = 1;
+         tileMap->boothCount = 2;
+         tileMap->booths[0].id = 5;
          tileMap->booths[0].tileIndex = 198;
+         tileMap->booths[1].id = 0;
+         tileMap->booths[1].tileIndex = 714;
          tileMap->portals[0].sourceTileIndex = 448;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 6777;
@@ -19775,8 +19765,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 187;
          tileMap->portals[0].destinationTileMapIndex = 4;
          tileMap->portals[0].destinationTileIndex = 278;
@@ -19824,8 +19812,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 108;
          tileMap->portals[0].destinationTileMapIndex = 4;
          tileMap->portals[0].destinationTileIndex = 759;
@@ -19877,11 +19863,11 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 22;
          tileMap->tilesY = 22;
          tileMap->portalCount = 27;
-         tileMap->innTileIndex = 369;
-         tileMap->innPrice = 25;
-         tileMap->boothCount = 1;
-         tileMap->booths[0].id = 2;
+         tileMap->boothCount = 2;
+         tileMap->booths[0].id = 6;
          tileMap->booths[0].tileIndex = 407;
+         tileMap->booths[1].id = 1;
+         tileMap->booths[1].tileIndex = 369;
          tileMap->portals[0].sourceTileIndex = 308;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 1271;
@@ -20250,8 +20236,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 217;
          tileMap->portals[0].destinationTileMapIndex = 7;
          tileMap->portals[0].destinationTileIndex = 282;
@@ -20404,8 +20388,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 22;
          tileMap->tilesY = 22;
          tileMap->portalCount = 36;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 66;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 13474;
@@ -20753,15 +20735,15 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 30;
          tileMap->tilesY = 31;
          tileMap->portalCount = 72;
-         tileMap->innTileIndex = 128;
-         tileMap->innPrice = 100;
-         tileMap->boothCount = 3;
-         tileMap->booths[0].id = 4;
+         tileMap->boothCount = 4;
+         tileMap->booths[0].id = 8;
          tileMap->booths[0].tileIndex = 171;
-         tileMap->booths[1].id = 5;
+         tileMap->booths[1].id = 9;
          tileMap->booths[1].tileIndex = 385;
-         tileMap->booths[2].id = 6;
+         tileMap->booths[2].id = 10;
          tileMap->booths[2].tileIndex = 806;
+         tileMap->booths[3].id = 3;
+         tileMap->booths[3].tileIndex = 128;
          tileMap->portals[0].sourceTileIndex = 120;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 15342;
@@ -21558,8 +21540,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 30;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 25;
          tileMap->portals[0].destinationTileMapIndex = 10;
          tileMap->portals[0].destinationTileIndex = 340;
@@ -21776,11 +21756,11 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 26;
          tileMap->tilesY = 26;
          tileMap->portalCount = 47;
-         tileMap->innTileIndex = 124;
-         tileMap->innPrice = 20;
-         tileMap->boothCount = 1;
-         tileMap->booths[0].id = 3;
+         tileMap->boothCount = 2;
+         tileMap->booths[0].id = 7;
          tileMap->booths[0].tileIndex = 360;
+         tileMap->booths[1].id = 2;
+         tileMap->booths[1].tileIndex = 124;
          tileMap->portals[0].sourceTileIndex = 668;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 2634;
@@ -22380,11 +22360,9 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 32;
          tileMap->portalCount = 103;
-         tileMap->innTileIndex = 626;
-         tileMap->innPrice = 55;
          tileMap->boothCount = 1;
-         tileMap->booths[0].id = 7;
-         tileMap->booths[0].tileIndex = 312;
+         tileMap->booths[0].id = 4;
+         tileMap->booths[0].tileIndex = 626;
          tileMap->portals[0].sourceTileIndex = 32;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 11171;
@@ -23277,8 +23255,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 110;
          tileMap->portals[0].destinationTileMapIndex = 13;
          tileMap->portals[0].destinationTileIndex = 132;
@@ -23333,8 +23309,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 72;
          tileMap->portals[0].destinationTileMapIndex = 13;
          tileMap->portals[0].destinationTileIndex = 619;
@@ -23456,8 +23430,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 299;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 2698;
@@ -23522,8 +23494,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 595;
          tileMap->portals[0].destinationTileMapIndex = 16;
          tileMap->portals[0].destinationTileIndex = 596;
@@ -23586,8 +23556,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 36;
          tileMap->tilesY = 32;
          tileMap->portalCount = 4;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 587;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 8999;
@@ -23692,8 +23660,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 36;
          tileMap->tilesY = 32;
          tileMap->portalCount = 3;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 335;
          tileMap->portals[0].destinationTileMapIndex = 18;
          tileMap->portals[0].destinationTileIndex = 335;
@@ -23799,8 +23765,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 28;
          tileMap->tilesY = 49;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 263;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 7254;
@@ -23907,8 +23871,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 42;
          tileMap->tilesY = 38;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 857;
          tileMap->portals[0].destinationTileMapIndex = 7;
          tileMap->portals[0].destinationTileIndex = 42;
@@ -24053,8 +24015,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 36;
          tileMap->tilesY = 30;
          tileMap->portalCount = 6;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 383;
          tileMap->portals[0].destinationTileMapIndex = 23;
          tileMap->portals[0].destinationTileIndex = 449;
@@ -24158,8 +24118,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 42;
          tileMap->tilesY = 38;
          tileMap->portalCount = 7;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 449;
          tileMap->portals[0].destinationTileMapIndex = 22;
          tileMap->portals[0].destinationTileIndex = 383;
@@ -24338,8 +24296,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 27;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 427;
          tileMap->portals[0].destinationTileMapIndex = 23;
          tileMap->portals[0].destinationTileIndex = 608;
@@ -24392,8 +24348,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 22;
          tileMap->tilesY = 22;
          tileMap->portalCount = 4;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 472;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 7898;
@@ -24604,8 +24558,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 42;
          tileMap->tilesY = 38;
          tileMap->portalCount = 10;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 398;
          tileMap->portals[0].destinationTileMapIndex = 25;
          tileMap->portals[0].destinationTileIndex = 55;
@@ -24821,8 +24773,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 4;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 339;
          tileMap->portals[0].destinationTileMapIndex = 30;
          tileMap->portals[0].destinationTileIndex = 303;
@@ -24901,8 +24851,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 6;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 306;
          tileMap->portals[0].destinationTileMapIndex = 29;
          tileMap->portals[0].destinationTileIndex = 302;
@@ -24998,8 +24946,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 11;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 299;
          tileMap->portals[0].destinationTileMapIndex = 26;
          tileMap->portals[0].destinationTileIndex = 559;
@@ -25133,8 +25079,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 28;
          tileMap->portalCount = 4;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 299;
          tileMap->portals[0].destinationTileMapIndex = 32;
          tileMap->portals[0].destinationTileIndex = 363;
@@ -25203,8 +25147,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 23;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 363;
          tileMap->portals[0].destinationTileMapIndex = 30;
          tileMap->portals[0].destinationTileIndex = 464;
@@ -25238,8 +25180,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 23;
          tileMap->portalCount = 2;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 363;
          tileMap->portals[0].destinationTileMapIndex = 30;
          tileMap->portals[0].destinationTileIndex = 299;
@@ -25277,8 +25217,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 32;
          tileMap->tilesY = 32;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 971;
          tileMap->portals[0].destinationTileMapIndex = 32;
          tileMap->portals[0].destinationTileIndex = 372;
@@ -25597,8 +25535,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 249;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 1211;
@@ -25666,8 +25602,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 124;
          tileMap->portals[0].destinationTileMapIndex = 0;
          tileMap->portals[0].destinationTileIndex = 16358;
@@ -25735,8 +25669,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 129;
          tileMap->portals[0].destinationTileMapIndex = 10;
          tileMap->portals[0].destinationTileIndex = 64;
@@ -25763,8 +25695,6 @@ void TileMap_Load( TileMap_t* tileMap, uint32_t id )
          tileMap->tilesX = 20;
          tileMap->tilesY = 15;
          tileMap->portalCount = 1;
-         tileMap->innTileIndex = 0;
-         tileMap->innPrice = 0;
          tileMap->portals[0].sourceTileIndex = 148;
          tileMap->portals[0].destinationTileMapIndex = 10;
          tileMap->portals[0].destinationTileIndex = 147;
