+++ DragonQuestinoEditor/DragonQuestinoEditor/ViewModels/TileMapViewModel.cs
@@ -71,20 +71,6 @@ namespace DragonQuestinoEditor.ViewModels
          set => SetProperty( ref _isDark, value );
       }
 
-      private int _innTileIndex;
-      public int InnTileIndex
-      {
-         get => _innTileIndex;
-         set => SetProperty( ref _innTileIndex, value );
-      }
-
-      private int _innPrice;
-      public int InnPrice
-      {
-         get => _innPrice;
-         set => SetProperty( ref _innPrice, value );
-      }
-
       public TileMapViewModel( TileSet tileSet,
                                int id,
                                string? name,
@@ -94,8 +80,6 @@ namespace DragonQuestinoEditor.ViewModels
                                bool isDark,
                                int tilesX,
                                int tilesY,
-                               int innTileIndex,
-                               int innPrice,
                                int defaultTileTextureIndex )
       {
          _id = id;
@@ -106,8 +90,6 @@ namespace DragonQuestinoEditor.ViewModels
          _isDark = isDark;
          _tilesX = tilesX;
          _tilesY = tilesY;
-         _innTileIndex = innTileIndex;
-         _innPrice = innPrice;
          EvacPortal = new( 0, 0, 0, Direction.Left );
 
          for ( int i = 0; i < tilesX * tilesY; i++ )
@@ -126,8 +108,6 @@ namespace DragonQuestinoEditor.ViewModels
          _isDark = saveData.IsDark;
          _tilesX = saveData.TilesX;
          _tilesY = saveData.TilesY;
-         _innTileIndex = saveData.InnTileIndex;
-         _innPrice = saveData.InnPrice;
          EvacPortal = new( saveData.EvacPortal );
 
          for ( int i = 0; i < saveData.Tiles.Count; i++ )
